ommit notes:  - Fixed issue with not showing failure when posting to Instagram  - Fixed issues with displaying Advocacy level  - Profile section hides when not logged in  - Removed tap through on feed  - Adjusted layout spacing on feed  - Now shows default image if noe profile image is present  - Fixed text on connect account dialog  - Nicer loading dialog on multi social login  - Fixed time display and spacing on rewards cards  - Fixed feed flashing on refresh  - After login will now go to profile screen  - UI alteratrion to the rewards hitory  - Fixed issue with being able to press on a reward in the rewards fragment multiple times

+++ sdk/src/main/java/com/popdeem/sdk/core/realm/PDRealmUserDetails.java
@@ -65,7 +65,7 @@ public class PDRealmUserDetails extends RealmObject {
         this.userFacebook = new PDRealmUserFacebook(user.getPdUserFacebook());
         this.userInstagram = new PDRealmUserInstagram(user.getPdUserInstagram());
         this.userTwitter = new PDRealmUserTwitter(user.getPdUserTwitter());
-        this.advocacy_score = user.getAdvocacyScore();
+        this.advocacy_score = Float.valueOf(user.getAdvocacyScore());
     }
 
     public int getUid() {
